+++ Newt/Models/DBRelationship.cs
@@ -1,0 +1,8 @@
+namespace Newt.Models
+{
+    internal class DBRelationship
+    {
+        public DBTable? Table { get; init; }
+        public DBConstraint? Constraint { get; set; }
+    }
+}